Inclass update
NOTICE: I changed the Hashmap to <String, Author>.  It was absolutely
useless to have it the other way and complicated the code.  The Author
class holds the papers that each author has written and can be accessed
by nameAuthorMap.get(author).getPaperList();

I have updated your read/write bin methods and read text method to
reflect this change.  This is untested but should work.  If not, try
and correct it or contact me if you're not sure what I did.